feat: add zora nft separation
