+++ apps/web/src/components/Publication/OpenActions/Nft/ZoraNft/Mint/Price.tsx
